Moved logo.svg to src/assests, renamed ModalList component
